Refactor Connection buttons
