+++ src/components/connection/Connection.vue
@@ -15,22 +15,12 @@
       icon="flight_land"
     )
     connection-settings
-    v-layout(row)
-      template(v-for="button in buttons")
-        v-flex(
-          :key="button.id"
-          :xs5="button.xs5"
-          :xs6="button.xs6"
-        )
-          v-btn(
-            :block="button.block"
-            :color="button.color"
-            :large="button.large"
-            :loading="button.loading"
-            :disabled="buttonDisabled"
-            @click="button.callback"
-          ) {{ button.text }}
-        v-spacer(v-if="button.id === buttons[0].id")
+    connection-button(
+      :buttonDisabled="buttonDisabled"
+      :buttonLoading="loading"
+      @submit="getTrip"
+      @reset="resetForm"
+      )
     v-alert(
       color="error"
       icon="warning"
@@ -47,6 +37,7 @@ import { mapGetters } from 'vuex'
 
 import API from '@/api'
 import DepartureSearch from '@/components/departures/DepartureSearch'
+import ConnectionButton from '@/components/connection/ConnectionButton'
 import ConnectionSettings from '@/components/connection/ConnectionSettings'
 import Timeline from '@/components/timeline/Timeline'
 
@@ -54,78 +45,43 @@ export default {
   name: 'Connection',
   components: {
     DepartureSearch,
+    ConnectionButton,
     ConnectionSettings,
     Timeline
   },
   mounted () {
-    if (this.settings.submitButtonOnRightSide) {
-      this.toggleButtonOrder()
-    }
     setTimeout(() => {
       if (this.settings.autoRetrieveConnection) {
         this.getTrip()
       }
     }, 1000)
   },
-  data: vm => ({
-    buttons: [
-      {
-        id: 1,
-        name: 'submit',
-        text: 'Submit',
-        color: 'primary',
-        block: true,
-        large: true,
-        xs5: false,
-        xs6: true,
-        callback: vm.getTrip
-      },
-      {
-        id: 2,
-        name: 'reset',
-        text: 'Reset',
-        color: 'normal',
-        block: true,
-        large: true,
-        xs5: true,
-        xs6: false,
-        callback: vm.resetForm
-      }
-    ],
+  data: () => ({
     loading: false,
     takemehome: false,
-    dialog: false,
     alert: false,
     errors: null,
-    search: {
-      start: null,
-      end: null
-    },
     trips: null,
     target: '#timeline'
   }),
-  watch: {
-    buttonOrder: function () {
-      this.buttons.reverse()
-    }
-  },
   computed: {
     ...mapGetters(['items', 'stations', 'stationValid', 'settings']),
-    buttonOrder () {
-      return this.settings.submitButtonOnRightSide
-    },
     buttonDisabled () {
       return this.stationValid || this.takemehome
     }
   },
   methods: {
+    toggleLoading () {
+      this.loading = !this.loading
+      this.takemehome = !this.takemehome
+    },
     resetForm () {
       this.$store.dispatch('RESET_FORM')
       this.trips = ''
     },
     getTrip () {
       this.$store.dispatch('TOGGLE_LOADING')
-      this.takemehome = true
+      this.toggleLoading()
       this.trips = null
       this.errors = null
       this.alert = false
@@ -138,9 +94,9 @@ export default {
       })
         .then(response => {
           this.$store.dispatch('TOGGLE_LOADING')
-          this.takemehome = false
           this.trips = response.data.Trip
           this.$nextTick(() => {
+            this.toggleLoading()
             this.$vuetify.goTo(this.target, {
               offset: -30,
               duration: 700,
@@ -150,7 +106,7 @@ export default {
         })
         .catch(e => {
           this.$store.dispatch('TOGGLE_LOADING')
-          this.takemehome = false
+          this.toggleLoading()
           this.alert = true
           this.errors = 'Something went wrong with the API: "' + e + '".'
         })
